backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
python-dotenv==1.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
requests>=2.31.0
